Migrate Word Cloud visualization to React
